+++ gsi-server/src/test/java/de/gematik/idp/gsi/server/controller/FedIdpControllerTest.java
@@ -18,9 +18,10 @@ package de.gematik.idp.gsi.server.controller;
 
 import static de.gematik.idp.IdpConstants.FED_AUTH_ENDPOINT;
 import static de.gematik.idp.IdpConstants.TOKEN_ENDPOINT;
+import static de.gematik.idp.data.Oauth2ErrorCode.UNAUTHORIZED_CLIENT;
 import static de.gematik.idp.gsi.server.data.GsiConstants.FEDIDP_PAR_AUTH_ENDPOINT;
 import static de.gematik.idp.gsi.server.data.GsiConstants.FED_SIGNED_JWKS_ENDPOINT;
-import static de.gematik.idp.gsi.server.data.GsiConstants.REQUEST_URI_TTL_SECS;
+import static de.gematik.idp.gsi.server.data.GsiConstants.TLS_CLIENT_CERT_HEADER_NAME;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 
@@ -29,25 +30,30 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import de.gematik.idp.IdpConstants;
 import de.gematik.idp.authentication.UriUtils;
+import de.gematik.idp.crypto.CryptoLoader;
 import de.gematik.idp.field.ClientUtilities;
 import de.gematik.idp.field.CodeChallengeMethod;
 import de.gematik.idp.gsi.server.GsiServer;
+import de.gematik.idp.gsi.server.configuration.GsiConfiguration;
 import de.gematik.idp.gsi.server.data.ClaimsResponse;
 import de.gematik.idp.gsi.server.services.EntityStatementRpService;
 import de.gematik.idp.token.IdpJwe;
 import de.gematik.idp.token.JsonWebToken;
+import java.io.File;
+import java.nio.charset.StandardCharsets;
+import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
-import kong.unirest.HttpResponse;
-import kong.unirest.HttpStatus;
-import kong.unirest.JsonNode;
-import kong.unirest.Unirest;
+import kong.unirest.core.HttpResponse;
+import kong.unirest.core.HttpStatus;
+import kong.unirest.core.JsonNode;
+import kong.unirest.core.Unirest;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.http.HttpHeaders;
+import org.apache.commons.io.FileUtils;
 import org.awaitility.Awaitility;
 import org.jose4j.jwk.PublicJsonWebKey;
 import org.junit.jupiter.api.BeforeAll;
@@ -62,10 +68,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
 
 @Slf4j
 @ActiveProfiles("test-controller")
@@ -73,6 +82,11 @@ import org.springframework.test.context.ActiveProfiles;
 @DirtiesContext(classMode = ClassMode.AFTER_CLASS)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class FedIdpControllerTest {
+
+  @DynamicPropertySource
+  static void dynamicProperties(final DynamicPropertyRegistry registry) {
+    registry.add("gsi.requestUriTTL", () -> 5);
+  }
 
   @Autowired private EntityStatementRpService entityStatementRpService;
   static final List<String> OPENID_PROVIDER_CLAIMS =
@@ -96,7 +110,9 @@ class FedIdpControllerTest {
           "id_token_signing_alg_values_supported",
           "id_token_encryption_alg_values_supported",
           "id_token_encryption_enc_values_supported",
-          "user_type_supported");
+          "user_type_supported",
+          "claims_supported",
+          "claims_parameter_supported");
 
   @LocalServerPort private int serverPort;
 
@@ -108,16 +124,32 @@ class FedIdpControllerTest {
   private HttpResponse<String> sigendJwksResponseGood;
   private JsonWebToken sigendJwks;
   private Map<String, Object> entityStatementbodyClaims;
+  @Autowired private GsiConfiguration gsiConfiguration;
+  private final String certFromRequest =
+      "-----BEGIN%20CERTIFICATE-----%0AMIIDszCCApugAwIBAgIUY%2FqefKABeWr36nT%2Brw9hJsbYFu8wDQYJKoZIhvcNAQEL%0ABQAwdjELMAkGA1UEBhMCREUxDzANBgNVBAgMBkJlcmxpbjEPMA0GA1UEBwwGQmVy%0AbGluMRkwFwYDVQQKDBBnZW1hdGlrVEVTVC1PTkxZMQ8wDQYDVQQLDAZQVCBJRE0x%0AGTAXBgNVBAMMEGZhZGlUbHNDbGllbnRSc2EwHhcNMjQwNjEzMDcxNjUyWhcNMjUw%0ANjEzMDcxNjUyWjB2MQswCQYDVQQGEwJERTEPMA0GA1UECAwGQmVybGluMQ8wDQYD%0AVQQHDAZCZXJsaW4xGTAXBgNVBAoMEGdlbWF0aWtURVNULU9OTFkxDzANBgNVBAsM%0ABlBUIElETTEZMBcGA1UEAwwQZmFkaVRsc0NsaWVudFJzYTCCASIwDQYJKoZIhvcN%0AAQEBBQADggEPADCCAQoCggEBAKiQaMTyY%2FlTTO9V4YJq7xsfN8l0%2BSqe2rRRasVU%0A8wenG8eohk99d1i5%2Fh08%2B%2BK1A5FX9GxgWh0RXGotpvbVvM7kzdOWxBJIK7j68R9g%0A%2F6B%2BKKO89rywLiJkxRT%2BOA4dusqocGDKmqFYZC1ntt2nSsSLlX3OuDC%2F1Thlhz2i%0AEGtweuYRL3zPeDXiegdyjRCY%2F9Xe%2FwaC4amuuJ5JkE5EsM0mL09kfkZCzdx8j2KK%0AqYTH2TYmiOG16CIVyZi9pE%2BKEHw95MIIcrzrO6QLWXcl7Y82rwVeeoUSicLBEydd%0A4YmsZ6pp%2BKGH0b9ycQO%2Bxs2uv79%2B5Zza9Q4OazEka4N0LyMCAwEAAaM5MDcwCQYD%0AVR0TBAIwADALBgNVHQ8EBAMCBeAwHQYDVR0OBBYEFMmogwgia7kONxur5UWBDX5g%0ABP0HMA0GCSqGSIb3DQEBCwUAA4IBAQAFK6nct1YVLMR6Tznh6ZrsvYs0UzCElUGM%0AnJtYaeCTgQPVKigQC4SPf%2FJp9qychooSbS7gbponndXgGIz8VFmt9y4d4q0uZKOr%0ALp7qcK%2BgQdvBts5TDZH20IiwW5b6VyGp%2Fos8fqR8WIt7fHdNz6Mu1fh2HsB4YjV9%0AxbbXTcKSzS6TROzh9bt2ubFX4ex56j6Mniy3DNF6zsW4kdh7naB%2FLfXvtH276Gj%2B%0AInhaF1sBLI8IIyQ5K2q2MJaly%2F8wiOys7FuG7duD1Lmh2kRO0FZkXsaQJmbZncUs%0A%2B4tgmnpEVgZ0FlKQ1BDAl0o0e7QbVRMiI2gjz7itOWFiUXvnMNIA%0A-----END%20CERTIFICATE-----%0A";
+  private X509Certificate certFromEntityStmtRpService;
+  private X509Certificate otherCert;
 
+  @SneakyThrows
   @BeforeAll
   void setup() {
     testHostUrl = "http://localhost:" + serverPort;
+
     entityStatementResponseGood = retrieveEntityStatement();
     sigendJwksResponseGood = retrieveSignedJwks();
     assertThat(entityStatementResponseGood.getStatus()).isEqualTo(HttpStatus.OK);
     entityStatement = new JsonWebToken(entityStatementResponseGood.getBody());
     sigendJwks = new JsonWebToken(sigendJwksResponseGood.getBody());
     entityStatementbodyClaims = entityStatement.extractBodyClaims();
+
+    certFromEntityStmtRpService =
+        CryptoLoader.getCertificateFromPem(
+            java.net.URLDecoder.decode(certFromRequest, StandardCharsets.UTF_8).getBytes());
+
+    otherCert =
+        CryptoLoader.getCertificateFromPem(
+            FileUtils.readFileToByteArray(
+                new File("src/test/resources/keys/fachdienstTlsC_TEST_ONLY.pem")));
   }
 
   @BeforeEach
@@ -185,8 +217,10 @@ class FedIdpControllerTest {
     assertThat(openidProvider.get("logo_uri")).asString().isNotEmpty();
     assertThat(openidProvider)
         .containsEntry("authorization_endpoint", gsiServerUrl + "/auth")
-        .containsEntry("token_endpoint", gsiServerUrl + "/token")
-        .containsEntry("pushed_authorization_request_endpoint", gsiServerUrl + "/PAR_Auth");
+        .containsEntry("token_endpoint", gsiConfiguration.getServerUrlMtls() + "/token")
+        .containsEntry(
+            "pushed_authorization_request_endpoint",
+            gsiConfiguration.getServerUrlMtls() + "/PAR_Auth");
     assertThat((List) openidProvider.get("client_registration_types_supported"))
         .containsExactlyInAnyOrder("automatic");
     assertThat((List) openidProvider.get("subject_types_supported"))
@@ -585,7 +619,7 @@ class FedIdpControllerTest {
     final String requestUri =
         ((JsonObject) JsonParser.parseString(respMsg3.getBody())).get("request_uri").getAsString();
 
-    waitForSeconds(REQUEST_URI_TTL_SECS + 2);
+    waitForSeconds(gsiConfiguration.getRequestUriTTL() + 2);
 
     Unirest.config().reset().followRedirects(false);
     final HttpResponse<String> respMsg6a =
@@ -608,7 +642,7 @@ class FedIdpControllerTest {
             .field("client_id", "https://DUMMY_CLIENT.de")
             .field("redirect_uri", "DUMMY_REDIRECT_URI")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .header(org.springframework.http.HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .asJson();
     assertThat(httpResponse.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
   }
@@ -623,7 +657,7 @@ class FedIdpControllerTest {
             .field("client_id", "https://DUMMY_CLIENT.de")
             .field("redirect_uri", "DUMMY_REDIRECT_URI")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .header(org.springframework.http.HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .asJson();
     assertThat(httpResponse.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
   }
@@ -704,8 +738,8 @@ class FedIdpControllerTest {
             .field("client_id", fachdienstClientId)
             .field("redirect_uri", redirectUri)
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .header(jakarta.ws.rs.core.HttpHeaders.USER_AGENT, "IdP-Client")
-            .header(org.springframework.http.HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.USER_AGENT, "IdP-Client")
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .asJson();
     assertThat(httpResponse.getStatus()).isEqualTo(HttpStatus.OK);
 
@@ -786,8 +820,8 @@ class FedIdpControllerTest {
             .field("client_id", fachdienstClientId)
             .field("redirect_uri", redirectUri)
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .header(jakarta.ws.rs.core.HttpHeaders.USER_AGENT, "IdP-Client")
-            .header(org.springframework.http.HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.USER_AGENT, "IdP-Client")
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .asJson();
     assertThat(httpResponse.getStatus()).isEqualTo(HttpStatus.OK);
 
@@ -796,6 +830,96 @@ class FedIdpControllerTest {
 
     // verify that token is encrypted and check kid
     assertThat(idpJwe.extractHeaderClaims()).containsEntry("kid", KEY_ID);
+  }
+
+  @Test
+  void parRequest_validAmr_httpsStatus_201() {
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<String> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .field("amr", "urn:telematik:auth:eGK")
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asString();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.CREATED);
+  }
+
+  @Test
+  void parRequest_invalidAmr_httpsStatus_400_message() {
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<JsonNode> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .field("amr", "urn:telematik:auth:invalid")
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asJson();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
+    assertThat(respMsg3.getBody().getObject().getString("error_description"))
+        .contains("amr: must match");
+  }
+
+  @Test
+  void parRequest_prompt_httpsStatus_201() {
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<JsonNode> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .field("prompt", "login")
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asJson();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.CREATED);
+  }
+
+  @Test
+  void parRequest_max_age_httpsStatus_201() {
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<JsonNode> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .field("max_age", "0")
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asJson();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.CREATED);
   }
 
   @SneakyThrows
@@ -883,6 +1007,35 @@ class FedIdpControllerTest {
     assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
   }
 
+  /**
+   * ClientID may contain just http and not https. This should work as well for local development
+   * environment.
+   */
+  @Test
+  void parRequest_authRequestUriPar_ClientIdHttp_httpStatus_201() {
+    Mockito.doNothing()
+        .when(entityStatementRpService)
+        .doAutoregistration(testHostUrl, testHostUrl + "/AS", "urn:telematik:given_name openid");
+
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+
+    final HttpResponse<String> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", "http://127.0.0.1:8084")
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:given_name openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asString();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.CREATED);
+  }
+
   /** Increase Test coverage of Token endpoint */
   @Test
   void
@@ -932,8 +1085,8 @@ class FedIdpControllerTest {
             .field("client_id", fachdienstClientId)
             .field("redirect_uri", redirectUri)
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .header(jakarta.ws.rs.core.HttpHeaders.USER_AGENT, "IdP-Client")
-            .header(org.springframework.http.HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.USER_AGENT, "IdP-Client")
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .asJson();
     assertThat(httpResponse1.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
 
@@ -945,10 +1098,174 @@ class FedIdpControllerTest {
             .field("client_id", fachdienstClientId)
             .field("redirect_uri", "invalidRedirectUri")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .header(jakarta.ws.rs.core.HttpHeaders.USER_AGENT, "IdP-Client")
-            .header(org.springframework.http.HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.USER_AGENT, "IdP-Client")
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
             .asJson();
     assertThat(httpResponse2.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST);
+  }
+
+  @SneakyThrows
+  @Test
+  void parRequest_tlsCertificateHeader() {
+
+    Mockito.doReturn(certFromEntityStmtRpService)
+        .when(entityStatementRpService)
+        .getRpTlsClientCert(any());
+
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<String> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .header(TLS_CLIENT_CERT_HEADER_NAME, certFromRequest)
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asString();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.CREATED);
+  }
+
+  @SneakyThrows
+  @Test
+  void parRequest_invalidTlsCertificateHeader_anyString() {
+
+    Mockito.doReturn(certFromEntityStmtRpService)
+        .when(entityStatementRpService)
+        .getRpTlsClientCert(any());
+
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<JsonNode> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .header(TLS_CLIENT_CERT_HEADER_NAME, "AnyInvalidString")
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asJson();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED);
+    assertThat(respMsg3.getBody().getObject().getString("error"))
+        .isEqualTo(UNAUTHORIZED_CLIENT.getSerializationValue());
+    assertThat(respMsg3.getBody().getObject().getString("error_description"))
+        .contains("client certificate in tls handshake is not a valid x509 certificate");
+  }
+
+  @SneakyThrows
+  @Test
+  void parRequest_certInHeaderAndInEntityStatementDontMatch() {
+
+    Mockito.doReturn(otherCert).when(entityStatementRpService).getRpTlsClientCert(any());
+
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<JsonNode> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .header(TLS_CLIENT_CERT_HEADER_NAME, certFromRequest)
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asJson();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.UNAUTHORIZED);
+    assertThat(respMsg3.getBody().getObject().getString("error"))
+        .isEqualTo(UNAUTHORIZED_CLIENT.getSerializationValue());
+    assertThat(respMsg3.getBody().getObject().getString("error_description"))
+        .contains(
+            "client certificate in tls handshake does not match certificate in entity"
+                + " statement/signed_jwks");
+  }
+
+  @SneakyThrows
+  @Test
+  void tokenRequest_tlsCertificateHeader() {
+    final String KEY_ID = "puk_fd_enc";
+    // key from gra-server/src/main/resources/keys/gras-enc-pubkey.pem
+    final String JWK_AS_STRING_PUK_FED_ENC =
+        "{\"use\": \"enc\",\"kid\": \""
+            + KEY_ID
+            + "\",\"kty\": \"EC\",\"crv\": \"P-256\",\"x\":"
+            + " \"NQLaWbuQDHgSHahqb9zxlDdiMCHXSgY0L9ql1k7BVUE\",\"y\":"
+            + " \"_USgmqhlM3pvabkZ2SS_YE2Q57tTs6pK9cE_uZB-u3c\"}";
+
+    Mockito.doNothing()
+        .when(entityStatementRpService)
+        .doAutoregistration(testHostUrl, testHostUrl + "/AS", "urn:telematik:versicherter openid");
+
+    Mockito.doReturn(PublicJsonWebKey.Factory.newPublicJwk(JWK_AS_STRING_PUK_FED_ENC))
+        .when(entityStatementRpService)
+        .getRpEncKey(any());
+
+    Mockito.doReturn(certFromEntityStmtRpService)
+        .when(entityStatementRpService)
+        .getRpTlsClientCert(any());
+
+    final String codeVerifier = ClientUtilities.generateCodeVerifier();
+    final String redirectUri = testHostUrl + "/AS";
+    final String fachdienstClientId = testHostUrl;
+    final HttpResponse<String> respMsg3 =
+        Unirest.post(testHostUrl + FEDIDP_PAR_AUTH_ENDPOINT)
+            .field("client_id", fachdienstClientId)
+            .field("state", "state_Fachdienst")
+            .field("redirect_uri", redirectUri)
+            .field("code_challenge", ClientUtilities.generateCodeChallenge(codeVerifier))
+            .field("code_challenge_method", CodeChallengeMethod.S256.toString())
+            .field("response_type", "code")
+            .field("nonce", "42")
+            .field("scope", "urn:telematik:versicherter openid")
+            .field("acr_values", "gematik-ehealth-loa-high")
+            .header(TLS_CLIENT_CERT_HEADER_NAME, certFromRequest)
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .asString();
+    assertThat(respMsg3.getStatus()).isEqualTo(HttpStatus.CREATED);
+    final String requestUri =
+        ((JsonObject) JsonParser.parseString(respMsg3.getBody())).get("request_uri").getAsString();
+
+    Unirest.config().reset().followRedirects(false);
+    final HttpResponse<String> respMsg7 =
+        Unirest.get(testHostUrl + FED_AUTH_ENDPOINT)
+            .queryString("request_uri", requestUri)
+            .queryString("user_id", "12345678")
+            .asString();
+
+    assertThat(respMsg7.getStatus()).isEqualTo(HttpStatus.FOUND);
+
+    final String code =
+        UriUtils.extractParameterValue(respMsg7.getHeaders().get("Location").get(0), "code");
+
+    final HttpResponse<JsonNode> httpResponse =
+        Unirest.post(testHostUrl + TOKEN_ENDPOINT)
+            .field("grant_type", "authorization_code")
+            .field("code", code)
+            .field("code_verifier", codeVerifier)
+            .field("client_id", fachdienstClientId)
+            .field("redirect_uri", redirectUri)
+            .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+            .header(HttpHeaders.USER_AGENT, "IdP-Client")
+            .header(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+            .header(TLS_CLIENT_CERT_HEADER_NAME, certFromRequest)
+            .asJson();
+    assertThat(httpResponse.getStatus()).isEqualTo(HttpStatus.OK);
   }
 
   private void waitForSeconds(final int seconds) {